remove provider, add datasource

+++ cachebank/src/main/java/net/jspiner/cachebank/BankConstant.java
@@ -10,7 +10,7 @@ public final class BankConstant {
 
     public static int DEFAULT_MEM_CACHE_SIZE = 4 * 1024 * 1024; // 4mb
     public static int DEFAULT_DISK_CACHE_SIZE = 10 * 1024 * 1024; // 10mb
-    public static int DEFAULT_CACHE_TIME = 5 * 60 * 1000; // 5 minute
+    public static long DEFAULT_CACHE_TIME = 5 * 60 * 1000; // 5 minute
     public static CacheMode DEFAULT_CACHE_MODE = CacheMode.ALL;
 
 }
